refator and bug fix

+++ src/foam/nanos/cm/CMResult.js
@@ -14,7 +14,7 @@ foam.CLASS({
       name: 'key'
     },
     {
-      class: 'Float',
+      class: 'Double',
       name: 'value'
     }
   ]
